COM-3684: manon/sophie's 1st review
