+++ src/app/Authentication/styles.ts
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
-import { WHITE, GREY } from '../../styles/colors';
-import { MARGIN, PADDING } from '../../styles/metrics';
-import { FIRA_SANS_BLACK, FIRA_SANS_REGULAR } from '../../styles/fonts';
+import { WHITE, GREY } from '@/styles/colors';
+import { MARGIN, PADDING } from '@/styles/metrics';
+import { FIRA_SANS_BLACK, FIRA_SANS_REGULAR } from '@/styles/fonts';
 
 const styles = StyleSheet.create({
   image: {
