+++ src/app/Home/Catalog/styles.ts
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
-import { FIRA_SANS_REGULAR } from '../../../styles/fonts';
-import { MAIN_MARGIN_LEFT, MARGIN } from '../../../styles/metrics';
+import { FIRA_SANS_REGULAR } from '@/styles/fonts';
+import { MAIN_MARGIN_LEFT, MARGIN } from '@/styles/metrics';
 
 const styles = (count = { color: '', background: '' }) => StyleSheet.create({
   container: {
