+++ src/app/Home/LearnerCourses/index.tsx
@@ -6,23 +6,23 @@ import { connect } from 'react-redux';
 import { useIsFocused, CompositeScreenProps } from '@react-navigation/native';
 import { StackScreenProps } from '@react-navigation/stack';
 import get from 'lodash/get';
-import Courses from '../../../api/courses';
-import SubPrograms from '../../../api/subPrograms';
-import NextStepCell from '../../../components/steps/NextStepCell';
-import ProgramCell from '../../../components/ProgramCell';
-import CoursesSection, { EVENT_SECTION } from '../../../components/CoursesSection';
-import HomeScreenFooter from '../../../components/HomeScreenFooter';
-import { getLoggedUserId } from '../../../store/main/selectors';
-import commonStyles from '../../../styles/common';
-import { RootBottomTabParamList, RootStackParamList } from '../../../types/NavigationType';
-import { CourseType, SubProgramType, SubProgramWithProgramType } from '../../../types/CourseTypes';
-import { NextSlotsStepType } from '../../../types/StepTypes';
-import { getCourseProgress, getTheoreticalDuration } from '../../../core/helpers/utils';
-import { LEARNER, PEDAGOGY } from '../../../core/data/constants';
+import Courses from '@/api/courses';
+import SubPrograms from '@/api/subPrograms';
+import NextStepCell from '@/components/steps/NextStepCell';
+import ProgramCell from '@/components/ProgramCell';
+import CoursesSection, { EVENT_SECTION } from '@/components/CoursesSection';
+import HomeScreenFooter from '@/components/HomeScreenFooter';
+import { getLoggedUserId } from '@/store/main/selectors';
+import commonStyles from '@/styles/common';
+import { RootBottomTabParamList, RootStackParamList } from '@/types/NavigationType';
+import { CourseType, SubProgramType, SubProgramWithProgramType } from '@/types/CourseTypes';
+import { NextSlotsStepType } from '@/types/StepTypes';
+import { getCourseProgress, getTheoreticalDuration } from '@/core/helpers/utils';
+import { LEARNER, PEDAGOGY } from '@/core/data/constants';
 import styles from './styles';
-import { formatNextSteps, getElearningSteps } from '../../../screens/courses/list/helper';
-import LearnerEmptyState from '../../../screens/courses/list/LearnerEmptyState';
-import { StateType } from '../../../types/store/StoreType';
+import { formatNextSteps, getElearningSteps } from '@/core/helpers/courses';
+import LearnerEmptyState from '@/components/EmptyState/LearnerEmptyState';
+import { StateType } from '@/types/store/StoreType';
 
 interface LearnerCoursesProps extends CompositeScreenProps<
 StackScreenProps<RootBottomTabParamList>,
