+++ src/app/Home/LearnerCourses/styles.ts
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
-import { MAIN_MARGIN_LEFT, MARGIN, PADDING, BORDER_WIDTH, BORDER_RADIUS } from '../../../styles/metrics';
-import { GREEN, GREY, PINK, PURPLE, YELLOW, TRANSPARENT_GREY, WHITE } from '../../../styles/colors';
-import { FIRA_SANS_REGULAR, FIRA_SANS_BOLD } from '../../../styles/fonts';
+import { MAIN_MARGIN_LEFT, MARGIN, PADDING, BORDER_WIDTH, BORDER_RADIUS } from '@/styles/metrics';
+import { GREEN, GREY, PINK, PURPLE, YELLOW, TRANSPARENT_GREY, WHITE } from '@/styles/colors';
+import { FIRA_SANS_REGULAR, FIRA_SANS_BOLD } from '@/styles/fonts';
 
 const styles = StyleSheet.create({
   container: {
