+++ src/app/Home/profile/PasswordEdition/index.tsx
@@ -1,9 +1,9 @@
 import { CompositeScreenProps } from '@react-navigation/native';
 import { StackScreenProps } from '@react-navigation/stack';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { RootStackParamList, RootBottomTabParamList } from '../../../../types/NavigationType';
-import PasswordForm from '../../../../components/PasswordForm';
-import Authentication from '../../../../api/authentication';
+import { RootStackParamList, RootBottomTabParamList } from '@/types/NavigationType';
+import PasswordForm from '@/components/PasswordForm';
+import Authentication from '@/api/authentication';
 import styles from './styles';
 
 interface PasswordEditionProps extends CompositeScreenProps<
