+++ src/app/Home/profile/Profile/index.tsx
@@ -7,34 +7,34 @@ import { connect } from 'react-redux';
 import get from 'lodash/get';
 import { StackScreenProps } from '@react-navigation/stack';
 import { CompositeScreenProps, useIsFocused } from '@react-navigation/native';
-import { PictureType } from '../../../../types/PictureTypes';
-import { RootBottomTabParamList, RootStackParamList } from '../../../../types/NavigationType';
-import { formatPhone, getCourseProgress } from '../../../../core/helpers/utils';
-import NiSecondaryButton from '../../../../components/form/SecondaryButton';
-import NiPrimaryButton from '../../../../components/form/PrimaryButton';
-import commonStyles from '../../../../styles/common';
-import { AuthContextType, Context as AuthContext } from '../../../../context/AuthContext';
-import Courses from '../../../../api/courses';
-import CompanyLinkRequests from '../../../../api/companyLinkRequests';
-import Users from '../../../../api/users';
-import Companies from '../../../../api/companies';
-import { PINK } from '../../../../styles/colors';
-import { UserType } from '../../../../types/UserType';
-import { HIT_SLOP, ICON } from '../../../../styles/metrics';
-import FeatherButton from '../../../../components/icons/FeatherButton';
-import PictureModal from '../../../../components/PictureModal';
-import CompanySearchModal from '../../../../components/CompanySearchModal';
-import DeletionConfirmationModal from '../../../../components/DeletionConfirmationModal';
-import UserAccountDeletedModal from '../../../../components/UserAccountDeletedModal';
-import HomeScreenFooter from '../../../../components/HomeScreenFooter';
-import CameraModal from '../../../../components/camera/CameraModal';
-import ImagePickerManager from '../../../../components/ImagePickerManager';
-import ValidationModal from '../../../../components/companyLinkRequest/ValidationModal';
-import { formatImage, formatPayload } from '../../../../core/helpers/pictures';
-import MainActions from '../../../../store/main/actions';
-import { PEDAGOGY } from '../../../../core/data/constants';
-import { ActionType, ActionWithoutPayloadType, StateType } from '../../../../types/store/StoreType';
-import { CompanyType } from '../../../../types/CompanyType';
+import { PictureType } from '@/types/PictureTypes';
+import { RootBottomTabParamList, RootStackParamList } from '@/types/NavigationType';
+import { formatPhone, getCourseProgress } from '@/core/helpers/utils';
+import NiSecondaryButton from '@/components/form/SecondaryButton';
+import NiPrimaryButton from '@/components/form/PrimaryButton';
+import commonStyles from '@/styles/common';
+import { AuthContextType, Context as AuthContext } from '@/context/AuthContext';
+import Courses from '@/api/courses';
+import CompanyLinkRequests from '@/api/companyLinkRequests';
+import Users from '@/api/users';
+import Companies from '@/api/companies';
+import { PINK } from '@/styles/colors';
+import { UserType } from '@/types/UserType';
+import { HIT_SLOP, ICON } from '@/styles/metrics';
+import FeatherButton from '@/components/icons/FeatherButton';
+import PictureModal from '@/components/PictureModal';
+import CompanySearchModal from '@/components/CompanySearchModal';
+import DeletionConfirmationModal from '@/components/DeletionConfirmationModal';
+import UserAccountDeletedModal from '@/components/UserAccountDeletedModal';
+import HomeScreenFooter from '@/components/HomeScreenFooter';
+import CameraModal from '@/components/camera/CameraModal';
+import ImagePickerManager from '@/components/ImagePickerManager';
+import ValidationModal from '@/components/companyLinkRequest/ValidationModal';
+import { formatImage, formatPayload } from '@/core/helpers/pictures';
+import MainActions from '@/store/main/actions';
+import { PEDAGOGY } from '@/core/data/constants';
+import { ActionType, ActionWithoutPayloadType, StateType } from '@/types/store/StoreType';
+import { CompanyType } from '@/types/CompanyType';
 import styles from './styles';
 
 interface ProfileProps extends CompositeScreenProps<
