+++ src/app/Home/profile/ProfileEdition/index.tsx
@@ -15,27 +15,27 @@ import { connect } from 'react-redux';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { StackScreenProps } from '@react-navigation/stack';
 import { CompositeScreenProps } from '@react-navigation/native';
-import FeatherButton from '../../../../components/icons/FeatherButton';
-import NiPrimaryButton from '../../../../components/form/PrimaryButton';
-import { GREY } from '../../../../styles/colors';
-import { ICON, IS_LARGE_SCREEN, MARGIN } from '../../../../styles/metrics';
-import { UserType } from '../../../../types/UserType';
-import { PictureType } from '../../../../types/PictureTypes';
+import FeatherButton from '@/components/icons/FeatherButton';
+import NiPrimaryButton from '@/components/form/PrimaryButton';
+import { GREY } from '@/styles/colors';
+import { ICON, IS_LARGE_SCREEN, MARGIN } from '@/styles/metrics';
+import { UserType } from '@/types/UserType';
+import { PictureType } from '@/types/PictureTypes';
 import styles from './styles';
-import NiInput from '../../../../components/form/Input';
-import { RootStackParamList, RootBottomTabParamList } from '../../../../types/NavigationType';
-import Users from '../../../../api/users';
-import { ActionType, ActionWithoutPayloadType, StateType } from '../../../../types/store/StoreType';
-import MainActions from '../../../../store/main/actions';
-import { EMAIL_REGEX, isIOS, PHONE_REGEX } from '../../../../core/data/constants';
-import ExitModal from '../../../../components/ExitModal';
-import NiErrorMessage from '../../../../components/ErrorMessage';
-import { formatPhoneForPayload } from '../../../../core/helpers/utils';
-import PictureModal from '../../../../components/PictureModal';
-import { errorReducer, initialErrorState, RESET_ERROR, SET_ERROR } from '../../../../reducers/error';
-import { formatImage, formatPayload } from '../../../../core/helpers/pictures';
-import CameraModal from '../../../../components/camera/CameraModal';
-import ImagePickerManager from '../../../../components/ImagePickerManager';
+import NiInput from '@/components/form/Input';
+import { RootStackParamList, RootBottomTabParamList } from '@/types/NavigationType';
+import Users from '@/api/users';
+import { ActionType, ActionWithoutPayloadType, StateType } from '@/types/store/StoreType';
+import MainActions from '@/store/main/actions';
+import { EMAIL_REGEX, isIOS, PHONE_REGEX } from '@/core/data/constants';
+import ExitModal from '@/components/ExitModal';
+import NiErrorMessage from '@/components/ErrorMessage';
+import { formatPhoneForPayload } from '@/core/helpers/utils';
+import PictureModal from '@/components/PictureModal';
+import { errorReducer, initialErrorState, RESET_ERROR, SET_ERROR } from '@/reducers/error';
+import { formatImage, formatPayload } from '@/core/helpers/pictures';
+import CameraModal from '@/components/camera/CameraModal';
+import ImagePickerManager from '@/components/ImagePickerManager';
 
 interface ProfileEditionProps extends CompositeScreenProps<
 StackScreenProps<RootStackParamList>,
