+++ src/components/EmptyState/LearnerEmptyState/index.tsx
@@ -1,6 +1,6 @@
 import { Text, Image, TouchableOpacity } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import NiPrimaryButton from '../../../../components/form/PrimaryButton';
+import NiPrimaryButton from '@/components/form/PrimaryButton';
 import styles from '../styles';
 
 const LearnerEmptyState = () => {
@@ -10,7 +10,7 @@ const LearnerEmptyState = () => {
     <TouchableOpacity style={styles.emptyStateContainer} onPress={() => navigation.navigate('Catalog')}>
       <Text style={styles.emptyStateText}>Vous n’avez pas de formation en cours...</Text>
       <NiPrimaryButton caption="Chercher une formation" onPress={() => navigation.navigate('Catalog')} />
-      <Image source={require('../../../../../assets/images/aux_detective.webp')} style={styles.emptyStateLearnerImage}
+      <Image source={require('../../../../assets/images/aux_detective.webp')} style={styles.emptyStateLearnerImage}
         resizeMode='contain' />
     </TouchableOpacity>
   );
